fn: get news list in latest news page
