+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+129",
+  "version": "0.8.5+130",
   "private": false,
   "scripts": {
     "dev": "next dev",
