+++ src/components/beta/dashboard/latest_news.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'next-i18next';
 import DashboardCardLayout from '@/components/beta/dashboard/dashboard_card_layout';
 import MoreLink from '@/components/beta/dashboard/more_link';
 import { ISUNFA_ROUTE } from '@/constants/url';
-import TabsForLatestNews from '@/components/beta/latest_news_page/tabs_for_latest_news';
-import NewsList from '@/components/beta/latest_news_page/news_list';
+import TabsForLatestNews from '@/components/beta/news_page/tabs_for_latest_news';
+import NewsList from '@/components/beta/news_page/news_list';
 import APIHandler from '@/lib/utils/api_handler';
 import { APIName } from '@/constants/api_connection';
 import { INews } from '@/interfaces/news';
