+++ src/components/beta/news_page/financial_news_page_body.tsx
