+++ src/components/beta/news_page/news_list.tsx
