+++ src/components/beta/news_page/system_news_page_body.tsx
