+++ src/components/beta/news_page/tabs_for_latest_news.tsx
