+++ src/constants/url.ts
@@ -21,10 +21,10 @@ export const ISUNFA_ROUTE = {
   DASHBOARD: '/users/dashboard',
   MY_COMPANY_LIST_PAGE: '/users/my_company_list_page',
   TODO_LIST_PAGE: '/users/todo_list_page',
-  LATEST_NEWS_PAGE: '/beta/latest_news_page',
-  FINANCIAL_NEWS_PAGE: '/beta/latest_news_page/financial_news_page',
-  SYSTEM_NEWS_PAGE: '/beta/latest_news_page/system_news_page',
-  MATCH_NEWS_PAGE: '/beta/latest_news_page/match_news_page',
+  LATEST_NEWS_PAGE: '/users/news_page',
+  FINANCIAL_NEWS_PAGE: '/users/news_page/financial_news_page',
+  SYSTEM_NEWS_PAGE: '/users/news_page/system_news_page',
+  MATCH_NEWS_PAGE: '/users/news_page/match_news_page',
   BETA_VOUCHER_LIST: '/users/accounting/voucher_list',
   GENERAL_SETTING: '/users/setting/general',
   BALANCE_SHEET: '/users/reports/financials/balance_sheet',
