+++ src/pages/users/news_page/index.tsx
@@ -1,17 +1,13 @@
 import Head from 'next/head';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import { GetServerSideProps } from 'next';
+import { ILocale } from '@/interfaces/locale';
 import { useTranslation } from 'next-i18next';
 import Layout from '@/components/beta/layout/layout';
 import { ISUNFA_ROUTE } from '@/constants/url';
-import SystemNewsPageBody from '@/components/beta/latest_news_page/system_news_page_body';
+import LatestNewsPageBody from '@/components/beta/news_page/news_page_body';
 
-interface SystemNewsPageProps {
-  newsId: string;
-}
-
-const SystemNewsPage = ({ newsId }: SystemNewsPageProps) => {
-  const { t } = useTranslation(['common']);
+const LatestNewsPage = () => {
+  const { t } = useTranslation(['dashboard']);
 
   return (
     <>
@@ -19,7 +15,7 @@ const SystemNewsPage = ({ newsId }: SystemNewsPageProps) => {
         <meta charSet="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon/favicon.ico" />
-        <title>{t('common:LATEST_NEWS_PAGE.SYSTEM_NEWS')}</title>
+        <title>{t('dashboard:LATEST_NEWS_PAGE.LATEST_NEWS')}</title>
         <meta
           name="description"
           content="iSunFA: Blockchain AI Forensic Accounting and Auditing is where simplicity meets accuracy in the realm of financial investigations."
@@ -36,28 +32,21 @@ const SystemNewsPage = ({ newsId }: SystemNewsPageProps) => {
 
       <Layout
         isDashboard={false}
-        pageTitle={t('common:LATEST_NEWS_PAGE.SYSTEM_NEWS')}
-        goBackUrl={ISUNFA_ROUTE.LATEST_NEWS_PAGE}
+        pageTitle={t('dashboard:LATEST_NEWS_PAGE.LATEST_NEWS')}
+        goBackUrl={ISUNFA_ROUTE.DASHBOARD}
       >
-        <SystemNewsPageBody newsId={newsId} />
+        <LatestNewsPageBody />
       </Layout>
     </>
   );
 };
 
-export const getServerSideProps: GetServerSideProps = async ({ params, locale }) => {
-  if (!params || !params.newsId || typeof params.newsId !== 'string') {
-    return {
-      notFound: true,
-    };
-  }
-
+export const getServerSideProps = async ({ locale }: ILocale) => {
   return {
     props: {
-      newsId: params.newsId,
-      ...(await serverSideTranslations(locale as string, ['layout', 'common'])),
+      ...(await serverSideTranslations(locale as string, ['layout', 'dashboard', 'date_picker'])),
     },
   };
 };
 
-export default SystemNewsPage;
+export default LatestNewsPage;
